+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Issue-291: Update `earthacess` version to enable use of Earthdata token for
   authentication.
+* Modify the OLVIS1A script to accept an environment option on the command line.
+* Remove OBE `CMR_URL` and hardcoding of provider name from OLVIS1A script.
 
 ## v1.14.0 (2026-06-18)